+++ src/routing.rs
@@ -154,6 +154,12 @@ WeighedShortest{
 }
 ```
 
+For multi-stage topologies we may use
+```
+UpDown{
+	legend_name: "up/down routing",
+}
+```
 
 ## Operations
 
@@ -253,6 +259,7 @@ pub fn new_routing(arg: RoutingBuilderArgument) -> Box<dyn Routing>
 			"Mindless" => Box::new(Mindless::new(arg)),
 			"WeighedShortest" => Box::new(WeighedShortest::new(arg)),
 			"Stubborn" => Box::new(Stubborn::new(arg)),
+			"UpDown" => Box::new(UpDown::new(arg)),
 			_ => panic!("Unknown Routing {}",cv_name),
 		}
 	}
@@ -1263,6 +1270,120 @@ impl Stubborn
 		let routing=routing.expect("There were no routing");
 		Stubborn{
 			routing,
+		}
+	}
+}
+
+///Use a shortest up/down path from origin to destination.
+///The up/down paths are understood as provided by `Topology::up_down_distance`.
+#[derive(Debug)]
+pub struct UpDown
+{
+}
+
+impl Routing for UpDown
+{
+	fn next(&self, _routing_info:&RoutingInfo, topology:&dyn Topology, current_router:usize, target_server:usize, num_virtual_channels:usize, _rng: &RefCell<StdRng>) -> Vec<CandidateEgress>
+	{
+		let (target_location,_link_class)=topology.server_neighbour(target_server);
+		let target_router=match target_location
+		{
+			Location::RouterPort{router_index,router_port:_} =>router_index,
+			_ => panic!("The server is not attached to a router"),
+		};
+		let (up_distance, down_distance) = topology.up_down_distance(current_router,target_router).unwrap_or_else(||panic!("The topology does not provide an up/down path from {} to {}",current_router,target_router));
+		if up_distance + down_distance == 0
+		{
+			for i in 0..topology.ports(current_router)
+			{
+				//println!("{} -> {:?}",i,topology.neighbour(current_router,i));
+				if let (Location::ServerPort(server),_link_class)=topology.neighbour(current_router,i)
+				{
+					if server==target_server
+					{
+						//return (0..num_virtual_channels).map(|vc|(i,vc)).collect();
+						return (0..num_virtual_channels).map(|vc|CandidateEgress::new(i,vc)).collect();
+					}
+				}
+			}
+			unreachable!();
+		}
+		let num_ports=topology.ports(current_router);
+		let mut r=Vec::with_capacity(num_ports*num_virtual_channels);
+		for i in 0..num_ports
+		{
+			//println!("{} -> {:?}",i,topology.neighbour(current_router,i));
+			if let (Location::RouterPort{router_index,router_port:_},_link_class)=topology.neighbour(current_router,i)
+			{
+				if let Some((new_u, new_d)) = topology.up_down_distance(router_index,target_router)
+				{
+					if (new_u<up_distance && new_d<=down_distance) || (new_u<=up_distance && new_d<down_distance)
+					{
+						r.extend((0..num_virtual_channels).map(|vc|CandidateEgress::new(i,vc)));
+					}
+				}
+			}
+		}
+		//println!("From router {} to router {} distance={} cand={}",current_router,target_router,distance,r.len());
+		r
+	}
+	fn initialize_routing_info(&self, _routing_info:&RefCell<RoutingInfo>, _topology:&dyn Topology, _current_router:usize, _target_server:usize, _rng: &RefCell<StdRng>)
+	{
+	}
+	fn update_routing_info(&self, _routing_info:&RefCell<RoutingInfo>, _topology:&dyn Topology, _current_router:usize, _current_port:usize, _target_server:usize, _rng: &RefCell<StdRng>)
+	{
+	}
+	fn initialize(&mut self, _topology:&Box<dyn Topology>, _rng: &RefCell<StdRng>)
+	{
+	}
+	fn performed_request(&self, _requested:&CandidateEgress, _routing_info:&RefCell<RoutingInfo>, _topology:&dyn Topology, _current_router:usize, _target_server:usize, _num_virtual_channels:usize, _rng:&RefCell<StdRng>)
+	{
+	}
+	fn statistics(&self, _cycle:usize) -> Option<ConfigurationValue>
+	{
+		return None;
+	}
+	fn reset_statistics(&mut self, _next_cycle:usize)
+	{
+	}
+}
+
+impl UpDown
+{
+	pub fn new(arg: RoutingBuilderArgument) -> UpDown
+	{
+		//let mut order=None;
+		//let mut servers_per_router=None;
+		if let &ConfigurationValue::Object(ref cv_name, ref cv_pairs)=arg.cv
+		{
+			if cv_name!="UpDown"
+			{
+				panic!("A UpDown must be created from a `UpDown` object not `{}`",cv_name);
+			}
+			for &(ref name,ref _value) in cv_pairs
+			{
+				//match name.as_ref()
+				match AsRef::<str>::as_ref(&name)
+				{
+					//"order" => match value
+					//{
+					//	&ConfigurationValue::Array(ref a) => order=Some(a.iter().map(|v|match v{
+					//		&ConfigurationValue::Number(f) => f as usize,
+					//		_ => panic!("bad value in order"),
+					//	}).collect()),
+					//	_ => panic!("bad value for order"),
+					//}
+					"legend_name" => (),
+					_ => panic!("Nothing to do with field {} in UpDown",name),
+				}
+			}
+		}
+		else
+		{
+			panic!("Trying to create a UpDown from a non-Object");
+		}
+		//let order=order.expect("There were no order");
+		UpDown{
 		}
 	}
 }