+++ src/topology/neighbourslists.rs
@@ -150,6 +150,10 @@ impl Topology for NeighboursLists
 		//input_port/2 != output_port/2
 		true
 	}
+	fn up_down_distance(&self,_origin:usize,_destination:usize) -> Option<(usize,usize)>
+	{
+		None
+	}
 }
 
 impl NeighboursLists